IJPL-194822 Added xi:include path resolver
All content module descriptors are inlined into plugin descriptor and loaded at once. Later, they are parsed using this branch, and missing xi:include resolver prevents from using includes in there. Probably we don't want this in the end and want to migrate to the v2 optional modules, but good for now.

GitOrigin-RevId: d75fb7b67df41d3ef3235daf87419640d06923f0