[gradle] fix identity path calculation for buildSrc of a root build (for Gradle < 8.0)
Before, `DefaultGradleLightProject#getBuildIdentityPath` did wrongly return `:` for this case, instead of `:buildSrc`.

However, identity path calculation for buildSrc of an included build is not supported yet. Instead of `:includedBuild:buildSrc` it returns just `:buildSrc`, because buildSrc is synced separately for Gradle < 8.0, so does not have a parent build.

Tests:
- org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test DefaultGradleLightProject.getProjectIdentityPath for buildSrc
- (FAILS FOR 7.6) org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test DefaultGradleLightProject.getProjectIdentityPath for buildSrc of included build

Is done as a previous step for IDEA-370426, but is far from the context described in the ticket.

GitOrigin-RevId: d5a8855da45c9a8399cc6e0ae69e4ae3cc978afa