[platform] move moduleExtension extension point from module-level to application level (IJPL-179175)
'ModuleExtension' was registered as a module-level interface extension point. Now it's an application-level bean-class extension point. It still supports calling the constructor with a 'Module' parameter, so compatibility with plugins isn't broken. Handling changes in the extension point is now performed in the app-level LegacyModuleExtensionRegistry service instead of registering a separate listener in each module.

GitOrigin-RevId: e71531ba38e9d89f67ee6ee5a67030210e70c232