[gradle] add test for buildTreePath usage as an identity path in Gradle >= 8.2 instead of calculating it
Also, create isBuildTreePathAvailable method in the GradleVersionSpecificsUtil object. It seems convenient to keep Gradle version specifics in one util object, to clearly display the difference between Gradle versions in one place.

Tests:
- org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test identity path is not calculated when buildTreePath is available

Is done as a previous step for IDEA-370426, but is far from the context described in the ticket.

GitOrigin-RevId: ded15045fcc65e8a5e9c999557b19c88ac921521