[toolwindows] IJPL-195147 Add split right and down actions that create the copy of the current tool window tab
We already have "Split and move right/down" actions, but they just move the selected tab to the new tabs group.
This action creates the copy of the selected tool window tab and moves it to the new tabs group.
The logic to create the copy of the tool window content depends on the particular tool window, so the extension point was added.

GitOrigin-RevId: 9223075bb77c21e06ca89f395f0a0af80714e2d2