[terminal] IJPL-193309 TerminalOutputModel.replaceContent
Add a new, more generic, API to TerminalOutputModel.
While we're at it, also add an agnostic TerminalOffset interface.

For the implementation, mostly reuse the existing logic in
doUpdateContent. The cursor position update is moved
towards the end, though, but it's not very important:
trimming only happens if the document became larger,
not smaller, updating the cursor position and trimming
is more or less orthogonal. But it makes more sense this way,
and in the future we may update trimming logic,
so it may trim more aggressively, for example,
and then it will be important.

The most complicated part is updating the highlighting model.
We'll need a lot of tests to ensure there are no off-by-one
errors with all those binary searches.

GitOrigin-RevId: 4edc5f1e07657a39c3db119549ebfa0074b683a1