[terminal] IJPL-193309 Add tests for the new TerminalOutputModel API
These tests might seem an overkill, but they're very useful.
For example, try replacing
if (i >= 0) i else -i - 2
with
if (i >= 0) i + 1 else -i - 2
in updateHighlightings() and see how only ONE test fails.

These tests could be coded manually, but the amount of code
would be HUGE then. So we implement a rather complicated
parser for a special markup syntax. And add a ton of self-tests
for the parser, too.

GitOrigin-RevId: 75cd07cb0024f4020e42c321761aa17a4be6c599