IJPL-196073 Update TextIcon bounds when font size/style change
This fixes IDE-side scaling (Zoom IDE) when the icon is already
displayed when the scaling factor changes. In this case,
the font instance is not changed (setFont is not called),
but the value returned by myFont.getSize() does change.

Just in case, check the style as well, though it seldom changes.

GitOrigin-RevId: 47ea90e718ed0d001c0f9b360c6466510b7450a9