KTIJ-33154 [kotlin] Consistently use cancelling processor in `getAllElements`
In cases when `SubKey` is a very small percentage of all possible results
collected by `processAllElements`, simply returning `true` from processing without
any cancellation checks can be problematic.

This change makes sure that `processor` with cancellation checks
is consistently used for every processed result.

The downside is that now we have to use unchecked cast
of a collection.

GitOrigin-RevId: d475d20b0d4df33f8aff7cf65eb7b79659b40cf9