PY-82486 Don't use PyTypingTypeProvider.getClassType to process union types of form "Type1 | Type2"
The problem itself is caused by the recently introduced flag
`PyTypingTypeProvider.Context.myComputeTypeParameterScope` needed to prevent
SOE while detecting the scope owner for a type parameter.

Namely, in the provided test case, while the scope owner for the type parameter
`f:T` is being calculated, we temporarily set this flag to false in
`PyTypingTypeProvider.getTypeParameterScope`.

While going through the type hints of other parameters of `f`,
looking for the first occurrence of T, we start processing `Box2[T] | None`.
Because `PyTypingTypeProvider.getClassType` used to precede `getBinaryExpressionType`,
it first tries to infer the type of the binary expression itself through general
`TypeEvalContext.getType` (this behavior is problematic, and to be revised),
i.e. the implicit call of `Box2[T].__or__`. It initiates collecting inherited
type parameters for the `Box2` class in `PyTypeChecker.unifyReceiver`.
Doing that, we eventually process the type parameter `Box:T`, but now
we don't detect its scope owner due to the flag being false
(so it's inferred as just `T`).
Then, this bogus scope-less type parameter is cached in
`PyTypingTypeProvider.getGenericSubstitutions`.
Later, when myComputeTypeParameterScope is reset back, and we infer
the result of the call `unbox(x)`, we cannot substitute this scopeless T with
`Box:T` properly mapped to str, so the type parameter remains as-is in the result type.

GitOrigin-RevId: d142f5179f19107acb625f6c4df61b148fdc7881