[JEWEL-874] Fix `reverseLayout` on non-lazy scrolling containers
This fixes the behaviour of `reverseLayout` on non-lazy scrolling
containers. We had forgotten to pass the flag value to the
`reverseScrolling` parameter of the `vertical`/`horizontalScroll`
modifiers. Lazy overloads do not provide their own scrolling modifier,
and thus worked fine.

I also took the opportunity to add a bunch of KDoc clarifying how the
scrollbars & styling works, and fixed the scrollable containers' KDoc,
which were wrong or imprecise in several points.