[terminal] IJPL-196070 Fix restoring terminal tabs with different configuration than the initial one
The initial `ShellStartupOptions` we receive from the client are modified in `LocalTerminalDirectRunner.configureStartupOptions`. If we persist the modified command text, then on restoring the tab, it will be passed to `LocalTerminalDirectRunner.configureStartupOptions` again, and the final result won't be the same as it was with the initial command text.
Let's persist the initial shell command we receive from frontend. This way we ensure that the shell command in the case of opening the new tab and restoring the existing one will be the same.


(cherry picked from commit 30792c99a74aeac33c801f6736d1a2a094858cbf)

IJ-CR-168720

GitOrigin-RevId: 190de87e9c5d9124659323e839ae38ff6068f80c