IJPL-196231 Fix launch(Once)OnShow for unconfined coroutines
Dispatchers.Unconfined doesn't immediately invoke
the coroutine if already running an unconfined coroutine.
If the current coroutine is blocked on dialog.show(),
it means that no unconfined jobs will be executed
while the dialog is showing.

To work around it, we replace Dispatchers.Unconfined
with the appropriate UI dispatcher that is guaranteed
to not dispatch if used with any() modality.

Implement the fix under a registry key,
disabled for now, so we can safely pick it
to the release branches to have this as a possible
manual fallback for the users who experience this issue.

Reviewed in IJ-CR-168711 for 252 and release.

(cherry picked from commit 3cfd1a9e46d12d942c7ab6f30f158b71cd7e52d8)

GitOrigin-RevId: e8bf1909be5a023e1c1b0fb9d3b433509483b40e