[driver] Fix broken connection in case of fast IDE restart
In JmxCallHandler#invoke we have a cache for currentConnector  and if it happens that we initialize it but the connection is broken, we won't recreate it until the initial connection is closed. So we need to perform reset of connector in case there is connection issue.


(cherry picked from commit b236dc988b2ba5ab186bd804cbc92e8fe9c13bc1)

IJ-CR-168941

GitOrigin-RevId: 43b5702e1143a3f6864455f027b6f904f0ad2852