[java-execution] IDEA-364759 fix slowOp in ApplicationConfiguration: store FQCN instead of binary class name
Now we resolve the runtime-qualified class/method name on demand, at `ApplicationConfiguration#getRunClass` level.

We want to keep the ability to run the configuration even if the IDE is in dumb mode,
so now `ApplicationConfiguration#getMainClass` uses alternative resolve (`DumbService#withAlternativeResolveEnabled`).
The test `ApplicationRunClassDumbModeTest` has been added to verify this.

Inside `ClassEditorField`, the field storing ConcurrentMap has been removed.
It was part of a previous – failed – attempt to fix the slowOp,
but in reality it only changed the place where the slowOp occurred.

GitOrigin-RevId: bade72877c02c1bba095c69f264d5f868d41f4d7