IJPL-196420 Fix NPEs in AbstractPopup
NPE 1: showInFocusCenter() would not check
whether the frame is null before calling getRootPane().
At this moment it can be null if we're on the remdev welcome
screen, because it's a backend thing (Lux), so it doesn't exist
on the frontend (except as a Lux helper, but findVisibleFrame
isn't aware of it).

NPE 2: getWndManager() is technically nullable,
but it was used without checks. Not that it relevant
in production code, but false positive NPE warnings
is the last thing we want in Java code.

NPE 3: SpeedSearchSupply.getSupply was passed
a potentially nullable component. It's not obvious
if it can ever be null in practice, but there's a null
check right above the same place, so we better be
consistent.


(cherry picked from commit e32434c6a83f841031b8d8aa7624907be22c8ed6)

IJ-CR-169007

GitOrigin-RevId: a756b5dedaae20f628ff9c7b51552be6a1eb87a7