[plugin model tests] include proper modules in the core classloader when checking dependencies of the frontend process
In production, only modules with 'loading=embedded' and their dependencies are added to the main classloader of the core plugin, so the test should do the same.

This revealed a problem with intellij.microservices.ui plugin: its main module isn't loaded in the frontend because dependencies are missing, so this problem is temporarily suppressed.

GitOrigin-RevId: 83554104de2f6967965d462c76e169d9a7c2f2e1