[gradle] IDEA-351763 delete GradleFindUsagesTest
This test was moved to the quarantine. For the last 100 runs, it has a 100% success rate. However, the test itself doesn't seem to be useful.

It has only one method - `testCompletionForVersionCatalogProperty`, but actually it does not verify completion. It creates a project with a version catalog and usage of its library in build.gradle. The test verifies usage, however, it does so strangely: it takes a target PsiElement in `build.gradle`, and searches for itself.

It would make more sense to test take a PsiElement in a version catalog and verify that it has usages in build.gradle. This behavior is already tested in GradleVersionCatalogsFindUsagesTest, so I don't see any obstacle to deleting GradleFindUsagesTest.kt.

GitOrigin-RevId: 0c15a3d635c986036eb37c78c83b5ff9078460c4