[terminal] IJPL-193309 Improve type-ahead logic
To avoid changing the line length, when typing,
only type if there's nothing beyond the cursor,
or if there are only spaces. And if there are spaces,
replace them, not insert.

The same for backspace, only perform type-ahead for it
when the rest of the line is empty and replace
instead of delete.

GitOrigin-RevId: 3c06dfefe53f7c29973efa84b094e24bc6d68d48