[JEWEL-877] Include CSS LinearGradientBrush and sample
This adds the `Brush.cssLinearGradientBrush()` API, which allows users
to draw linear gradients using a CSS-like semantic. This is based on a
blog post I wrote last year, and it's been used as-is already in the
Gemini in Android Studio UI.

In addition to the brush itself, this also adds a Brushes section to
the showcase where one can test and preview this brush' behaviour with
different parameters.

Lastly, this adds a few colour helper functions, including overloads
for existing ones (that are now deprecated in favour of the new ones).
This brings all colour util functions to support both ARGB and RGBA in
both input and output, to/from AWT and Compose `Color`s. We also now
finally have unit tests and proper KDoc for all the util functions.

Note: we're explicitly excluding the IDE sample plugin from the GitHub
CI as it's bound to have compilation issues as we add new APIs that
aren't yet available in a target build of the IDE. We still run ktlint
and ktfmt for the IDE sample, as they do not depend on compileKotlin.

closes https://github.com/JetBrains/intellij-community/pull/3121

GitOrigin-RevId: d512f23028d444e38a1018ea6e2ed0f44afd91f3