IJPL-184315 Add settings synchronization between frontend and backend
To not have conflicts on how the file looks on frontend and backend, the data consumer and provider for virtual files are registered on backend and frontend. Because settings are not working correctly if the settings provider class is on backend, the whole `intellij.sh.codeStyle` package was moved back to core. Service class was created to enable the usage of some backend classes.

GitOrigin-RevId: 951cd6df49432b5e93612b4d3c76452943426862