[JEWEL-891] Annotate detectPressAndCancel as internal API
It was opened up by someone without our knowledge, and it risks
leaking into the public API otherwise.

APIs annotated as @ApiStatus.Internal do not show up in the API dumps,
because as far as IJP is concerned, they should not be used. However,
we also have standalone users, and for them that annotation has no
meaning.

This will need to be cherry picked onto 252 asap, too.

closes https://github.com/JetBrains/intellij-community/pull/3122

GitOrigin-RevId: dfff8dcadd39cfbdc0e6557083bfa834cc1b9f47