[JEWEL-894] Fix EqualityMembersRule, add tests
The rule is a bit buggy, and difficult to read. It also
has essentially no tests. This addresses both concerns.
I also noticed for whatever reason we were not properly
running detektMain and detektTest on the GitHub CI,
which means we had a bunch of warnings. I have now added
both checks to the GitHub CI. Hopefully, we'll be able
to run the same checks on TeamCity at some point, too.

Note: this bumps Detekt to 1.23.8 (latest version) and
adds a test-only dependency to detekt-core — I'm not
sure if this will require manual merging due to Bazel.

It also tweaks the Jewel .editorconfig to 120 as that's
the value used by the CI, and fixes all outstanding
static analysis warnings.