[JEWEL-898] TooltipArea should not be public
Our public tooltip API should remain `Tooltip`, which delegates to `TooltipArea` that is a private impl detail. It should not be in the public API.

To avoid further confusion, the `TooltipArea` composable has also been renamed
to `FlagAwareTooltipArea`. This better describes its function and is clearly a
private API name.

This also fixes `ComposeShowcase` to use `Tooltip` instead of `TooltipArea`, and
sets up the Jewel VCS settings to recognise YouTrack issues. A few issues in the
`ComposeShowcase` have also been fixed.

closes https://github.com/JetBrains/intellij-community/pull/3127

GitOrigin-RevId: 9503539eab37161071c48f938ec0b4f9ff8464a5