[terminal] IJPL-195147 Fix dropping of files into the terminal when it is split
Dropping the file from the Project View should open the new terminal tab in the directory with this file.
Now it also creates the tab in the split area the file is dropped to.
We had to use the tool window decorator as the drop area to make it work. Because the component returned from `toolWindow.getComponent()` is not showing on the screen when the tool window is split.

GitOrigin-RevId: 53167bb3091b8cd6f61b2481230da4969fa9cf5a