[terminal] IJPL-195147 Fix dropping terminal tab moved to the editor back to the terminal tool window
Also, create the new tool window tab in the split area the editor tab is dropped to.
We had to use the tool window decorator as the drop area to make it work. Because the component returned from `toolWindow.getComponent()` is not showing on the screen when the tool window is split.

GitOrigin-RevId: 8803f278012939121c9453681731769ea0090bfc