[JEWEL-892] Add Detekt rule to check API annotations
This introduces a new Detekt rule, `MissingApiStatusAnnotationRule`,
that enforces the presence of matching ApiStatus annotations for
Jewel's API visibility annotations.

The rule ensures that:
* `@InternalJewelApi` is always paired with `@ApiStatus.Internal`
* `@ExperimentalJewelApi` is always paired with
  `@ApiStatus.Experimental`

This makes sure we do not mistakenly leak APIs to either the platform
(by only using Jewel annotations) or standalone (by only using the
ApiStatus annotations).

This also adds a few missing test cases from EqualityMembersRuleSpec,
and bumps Gradle's IJP target to 252.23892.248 (252 Beta 1).

Note: this also includes addressing all the issues that the new rule
flags (which is noisy), and adapting all the API dumps (which is uber
noisy) — sorry!