[terminal] IJPL-178962 Introduce a platform helper to add editor hyperlinks
The problem with EditorHyperlinkSupport is that it handles
too much stuff at once: hyperlink highlighting,
running filters, tracking document changes...
This doesn't fit well with scenarios like remdev,
when the document can exist without an editor on the frontend,
or specialized use cases when hyperlinks can be provided
by custom code instead of console filters,
or when document changes are restricted to some subset
(like in the terminal).

Let's start developing a new API, starting with just the UI
part that handles highlightings and clicks,
and the hyperlinks are provided by the clients.

GitOrigin-RevId: ae329edfa3accd738f85d2baee31f2a5c36f2a14