[JEWEL-906] Mandate valid JEWEL issue ID in single commit messages
This adds a check to GitHub Actions that validates:
 1. The PR has only one commit
 2. The commit message starts with "[JEWEL-xxx] ", where xxx is a
    positive number

These checks only run if the PR contains changes in the Jewel folder
other than `AllIconsKeys`. That gets autogenerated by a script when
any changes are made to IJP icons, and is high-traffic. Any IJP dev
can run the script, and it should not block PRs added elsewhere in the
IJP that update the icons.

Additionally, if there is a `JEWEL_YT_TOKEN` secret defined, the check
also validates that the referenced issue is valid — meaning, it
exists, is not resolved, and not a draft.