[JEWEL-899] Remove old Gradle plugins and tasks
This removes a number of now-unused Gradle plugins and tasks:
 * `ide-version-checker` (not used anymore)
 * `jewel-publish`, `PublishConfiguration`, and related tasks
   (publishing is handled by the JPS build)
 * `MergeSarifTask` (not used anymore) and related tasks
 * Related run configs

Note: `icon-keys-generator` has not been removed yet, even tho the
actual generator runs as part of the IJP's generation, since the
Gradle build would remain without the generated file otherwise. Since
we're planning to eventually sunset the Gradle build, this is only a
temporary necessity.

closes https://github.com/JetBrains/intellij-community/pull/3139

GitOrigin-RevId: e87697b81c7b581d6a5aa6d8daeeb18dbf622997