[platform] update all notifications on change in the plugin list
This should fix race conditions caused by attaching to a project disposable inside a constructor of the `EditorNotificationProvider` EP implementation.
The race condition roughly looks like this:
1. We start iterating over `point.sortedAdapters`
2. Check `coroutineContext.ensureActive()` succeeds
3. `adapter.createInstance<EditorNotificationProvider>` is called for some "bad" EP, thread suspends
4. Project was disposed in another thread
5. Thread wakes up, tries to attach its disposable to the project's disposable, but it fails

This issue would be better solved by throwing `AlreadyDisposedException` instead of `IncorrectOperationException` in the `ObjectTree` because ADE is a subclass of `CancellationException`, but this may break backwards compatibility.

GitOrigin-RevId: 041105e1180961508ea2aafebf0aa0a896d80a3a