[JEWEL-910] Fix GitHub CI to use the JBR
The CI stopped working seemingly overnight due to Gradle needing the
JetBrains Runtime, and the agents not having it. This makes the agents
use the JBR runtime instead of `zulu`, and updates the foojay plugin
to the latest version (as the plugin should be able to download the
JBR if missing anyway, so maybe that stopped working?).

This also bumps the Kotlin version in the Gradle build to match IJP
252, and replaces the deprecated Gradle Enterprise plugin with the new
Develocity one.

closes https://github.com/JetBrains/intellij-community/pull/3141

GitOrigin-RevId: d86c5e0cdb76ae83e95447b62eeefa9bd9b66014