[JEWEL-911] Fix infoForeground colour in Darcula theme
In the Darcula theme, for some reason, `infoForeground` is only
defined in `Component`, instead of being in the wildcard "*" node.
This results in the info text in this theme being displayed as normal
text.

This fixes the issue by moving all `BridgeGlobalColors` to use various
IJP known colour APIs instead of trying to read specific keys. These
APIs are "guaranteed" to work — or at least to break in the same ways
as Swing would.

It also adds two new colours to TextColors: disabledFocused and
warning.

closes https://github.com/JetBrains/intellij-community/pull/3144

GitOrigin-RevId: 2d4eb8315bf91e65f96354e3b1f94ed0d82a9d9b