KTIJ-31412 Disable K1/K2 mode switching for KotlinMppRunConfigurationsTest
API of overriding pluginMode(KTIJ-28619) was supposed to work only with one plugin mode. This problem is reproducible only on CI, where different versions of a test can be executed on one machine. For such a case the first setup of pluginMode will determine which mode will be loaded by KotlinPluginModeProvider, so if it is switched later, another service will not be loaded to classpath. This will cause `Invalid Kotlin plugin detected: K1, but K2 was expected expected:<K2> but was:<K1>` exception. To properly fix this problem, we need to create a separate test class with K2 mode

GitOrigin-RevId: 473c690d643b2a66e8f38935a1fcda280253941d