[JEWEL-916] Fix crash in UiInspectorInfo
On master, invoking the UI inspector when some (but not all!) Jewel UI
is attached can make the UI inspector crash. This is because our own
`UiInspectorInfo` is assuming all a11y related API return non-null
values, which clearly was a valid assumption until recently.

However, now some of those APIs do return null values, so we need to
treat them accordingly. This fixes the issue, and cleans up/simplifies
the code a bit in the process.