[JEWEL-917] Fix yet another crash due to malformed themes
As per https://issuetracker.google.com/433208919, there are malformed
themes out there that do not define Button.arc — which is considered a
mandatory key — and this causes crashes. This provides a defensive
fallback to account for these cases.

This also adds KDoc to BridgeUtils public API, a new
`Color.toAwtColorOrNull()` API, and rewrites some retrieval code to
avoid using try-catch (they're quite expensive), aligning all the
various retrieval impls to be based on the "or undefined" variants.