[JEWEL-918] Fix checkbox and radio button icons in Darcula
The Darcula theme is now considered "new UI". Our check to adopt the
"old themes" metrics (that consider that the Darcula SVGs are borked)
thus doesn't work any more. The fix is to explicitly adopt the "old
theme" metrics when the current theme is Darcula.

This only impacts the bridge, since we do not support Darcula in the
standalone mode.

closes https://github.com/JetBrains/intellij-community/pull/3148

GitOrigin-RevId: d12615b72984eeaae603e83df9d13bb1c7097b9d