[python][sdk] check poetry.lock in poetry auto configurator (PY-82840)
+ refactor PyProjectSdkConfiguration to return PyResult for SDK creation methods:
In case of creation failure, the actual error was hidden and some random sdk was assigned instead of the real one. For example, if we open a poetry project and project and poetry sdk creation fails - any random sdk was assigned to the project (last used hatch or conda for example). Now ShowMessageErrorSink is used.

GitOrigin-RevId: 015fc469b6b5d934b49278ccfb3afc683b61554e