[terminal] IJPL-178962 IncrementalUpdateFlowProducer: let the clients filter updates
The event filtering in StateAwareTerminalSession is a bit ugly
for several reasons. It exposes the underlying hyperlink
highlighter logic in isObsolete, and it doesn't prevent
the resulting empty lists from being forwarded to the client.

Let's fix this by improving IncrementalUpdateFlowProducer
to let the state modify updates. Then we can move filtering
to updateModelState.

GitOrigin-RevId: 9aab22b3baf4a0c3b6380c6f63e7628b1f5b93a6