[terminal] IJPL-178962 Hyperlinks: get rid of lists of events where not needed
The whole List<Event> thing exists only on the higher level,
and not even used for anything practical at the moment.
Let's not propagate this approach to lower level stuff
unless needed.

This might be reverted if we actually start to process EDT events
in bulks on the frontend side, then we can send several events
at once.

GitOrigin-RevId: 949dac5fb986be0a78c1a6afdd3008d984623f9f