[JEWEL-935] Make GitHub checks post comments
When there is some issue in a PR found by the GitHub actions workflow,
a comment is posted to the PR describing the issue found. This also
cleans up existing scripts and collects them in platform/jewel/scripts
so they do not pollute the rest of the codebase.

Lastly, this uses the GitHub CLI to count the commits in a PR as the
previous method had issues with rebases.

closes https://github.com/JetBrains/intellij-community/pull/3152

GitOrigin-RevId: e567e747ccae7a75064a242b1cb82d207c37ceff