IJPL-339 subscribe to context changes immeditely
This fixes com.intellij.java.codeInsight.AddAnnotationFixTest.testPrimitive test

The problem was that testPrimitive creates TestPrimitive.java and only after that adds a source root for the file.
And sometimes the event about source root was delivered earlier than EditorContextManagerImpl subscribed to them. And the editor context of Primitive.java was stuck with the default context, while the correct context was the module context

GitOrigin-RevId: 1d7f92a22597174a5c2013cfa26b87c8bdb74ad9