[terminal] IJPL-166283 Add shell integration to PowerShell
Thanks to the new ConPTY version, escape sequences output is now synchronized with text output, so now we have ordering guarantees.
Implemented the shell integration with the same principles as in Zsh and Bash.
It makes separators between the executed commands and error stripe show in the PowerShell as well.

GitOrigin-RevId: 7a44ece58d0f02c1457e7ce52f57d3b47992f7d8