[terminal] IJPL-199323 Fix showing error stripe for all commands after failed external app command
Let's do not try to capture the exit code of the externally launched applications. Since this value is persisted in the PowerShell variable, and we can't say to which command it relates.
We may reset it to zero after checking, but it means that we modify the state of the shell and might break something configured by the user. Better to avoid it.

GitOrigin-RevId: 21da7d9055e397b22a69938d5ccf5d67f7d56cae