[debugger] IDEA-343854 Extract `shouldKillProcessSoftly()` from `KillableProcessHandler` into an interface
`DebuggerManagerImpl` relies on the result of this method to decide if a JVM process should be stopped forcibly by terminating its VM or not. Initially, this method belonged to the `KillableProcessHandler` class. However, it could be useful for other process handlers, which processes should not be terminated by the debugger (e.g., ExternalSystemProcessHandler). So it was extracted into a separate interface.

GitOrigin-RevId: 8512892f9b8389afcf80ca16593f44094313a050