[JEWEL-936] Fix AlwaysVisible non-scrolling layout
In scrollable containers, when the `ScrollbarVisibility` is
`AlwaysVisible`, but the container is large enough that the content
doesn't scroll, we were still leaving empty space for the scrollbars,
which were invisible.

This fixes the issue by checking that the layout only reserves space
for the scrollbars when they can actually scroll, in `AlwaysVisible`.
This does not affect the `WhenScrolling` visibility setting, as that
does not reserve any space for the scrollbars anyway (they're overlaid
over the content).

This also improves the scrollbars showcase to have more representative
examples, and more comprehensive than before.

It also adds tests, a lot of tests! I even found and fixed a bug
thanks to the tests.

PS: I added a couple actions to the GitHub CI checks to annotate JUnit
test failures and upload the results, so they can be downloaded for
analysis, in case they only fail on CI.

PPS: I had to also fix an issue with AlwaysVisible and WhenScrolling
in standalone on Windows/Linux