[rdct] IJPL-170779: Fix settings synchronization for PersistentStateComponent with Element state
When the state of a PersistentStateComponent is represented by an Element, a single key is used for the whole state. If all settings in the PersistentStateComponent have default values, the initial value for the key in the RdSettingsStorageService is null. However, when the settings are saved, the current state is represented by an Element that contains only a tag name. This triggers synchronization since the Json representation of such Element isn't equal to null.

The fix treats the default value of a state represented by an Element as null.

GitOrigin-RevId: c83ee382715d93bd14220f74e321207e69e7e987