[terminal] IJPL-178962 The second split hyperlinks implementation
This change addresses the issues of the previous implementation.
It would cancel the entire task on change and redo a lot of work.
This new implementation doesn't cancel tasks. Instead,
it just tells the running task when to stop.
The results are always collected and validated.

Because only the model coroutine (the one that updates
the output model) can validate results without having
to worry about them becoming invalid the very next
millisecond, we have to do result collecting and task management
in the model coroutine now (the getResultsAndMaybeStartNewTask
function).

The tasks still run in background, and have isolated data.
Their only communications with the outside world are:

- continueCondition - checks if it makes sense to continue.
May return outdated results due to its async nature,
but it's designed to work on a best-effort basis:
if it didn't stop this time, it'll stop the next time.

- the predicate for validating results.
This one is only used when getNextOutputEvent
is called from the model coroutine,
so it has safe access to the latest state of the model.

To make this whole thing work with our event-based
approach, there's a fake heartbeat event now
which is never sent to the frontend,
and is only used to inform the event loop
that it's time to check for new results and/or manage tasks.

GitOrigin-RevId: cb269e150bb23bc0d217dfad475367d31073d79b