[JEWEL-946] Fix SimpleListItem styling
In the bridge, we were reading the simple list item colours from the
wrong keys. This was not obvious in the light theme, but in dark theme
this resulted in weird background colours. Additionally, we were not
correctly applying the "active" status to items in the popup, and the
inactive selection bg colour was incorrect in standalone, too.

This fixes both issues :)