[jps build] cleanup: deprecate JpsNamedElement::setName and remove its usages in tests
This is a followup for the introduction of JpsNamedElementCollection which was added to fix IDEA-368571. JpsNamedElementCollection doesn't handle renaming of elements. Given that JpsNamedElement::setName was used in two tests only, it's simpler to deprecate it and rewrite the test to remove and add modules instead of renaming them (this is how JPS build sees a change after a module is renamed in the IDE anyway).

GitOrigin-RevId: 98c151fb821c50465c90dab48f71ac8e1aaa58be