KTIJ-35047 [kotlin][lombok] Do not avoid wrapping added type parameters into `LightTypeParameter`
In the case when the generic class is used together with
`@Builder` annotation, its type parameter is added
to the generated `builder` method as is.

Simple example:
```
@Getter
@Builder
class Container<T> {
  private T value;
}
```

Because of this, later on Kotlin FIR confuses
the type parameter `T` of the generic class with the
type parameter of the `builder` method, which
messes up the resolve all occurrences of the
original `T` type in the class.

This is illustrated by the fact that:
1. The type parameter was resolved into something like `T! (of fun <T : Any!> builder)` (notice the `builder` here).
2. When the `@Builder` annotation was removed, the problem went away.

Add a highlighting test to check that scenario

^KTIJ-35047 Fixed

GitOrigin-RevId: f1909340b665cf71f09c84f40d255c1a68053882