PY-83039 Don't trigger PyLiteralTypeCompletionContributor in cases like `x = y.foo<caret>`
It should be activated only for the simplest cases when the caret is either
inside an immediate string literal or its prefix is an unqualified reference
expression. It makes little sense trying to detect if something like
the literal string `"y.foobar"` is a possible value for `x`.

It's a relatively heavy completion contributor. It starts evaluating the type of
`x` flow-sensitively, analyzing all preceding function calls to take into
account `NoReturn`. It affects common workflows like typing out
`df = pd.` to create a new Pandas dataframe.

GitOrigin-RevId: f17fe2ce86ee100a3480a574c7f57a1bd67ec2d8