[terminal] IJPL-178962 Implement hyperlink context menus
It's much more complicated than it would seem.
It doesn't even work in remdev yet because
data context serializers are missing.
They will be added separately, as it involves complicated
changes in the module system.

First, we need to pass the necessary IDs through the data context.
To identify a hyperlink, we need the session ID,
which buffer it's in, and the hyperlink ID.
To pass the session ID we need to expose the frontend session
as @Internal.

To actually access the link on the backend,
we need to get the session from the TerminalSessionsManager,
and then the hyperlink ID from the hyperlink backend.
To get the hyperlink facade, we need to expose StateAwareTerminalSession.

To make things worse, getPopupMenuGroup()
requires a non-nullable MouseEvent.
It's not used anywhere, but there are null checks
in external usages, so we can't just make it nullable.
For this to work we need to create a fake MouseEvent,
but it needs a component, and therefore it needs to be created
on the EDT. We delegate this job to the hyperlink engine.

GitOrigin-RevId: 727f399e491042c6f6851a8d49569a5030973d61