IJPL-184315 Fix build tests error caused by `intellij.sh.python` not being available on frontend
The package `intellij.sh.python` was unavailable on the frontend, and as frontend uses a module-based loader (builds a dependency graph from the runtime module repository), the entire plugin won't be loaded. The new module `intellij.sh.main` was introduced which aggregates all the dependencies of the Shell plugin modules that need to be included for IDEs runnning in any mode. Other `*.main` modules now depend on this module instead of `intellij.sh`.

GitOrigin-RevId: a8c357539806c45cdaa75d2420fae629e358698a