[rdct, rd-necropolis] IJPL-198448: Reuse opened cache whenever possible
If the frontend is reopened without stopping the client, it will try to register the remote caches on the backend. However, if the backend has not been stopped, it will leave the caches opened, so reregistration will fail with the exception.

This commit introduces the reusing of the caches, which should fix the exception

GitOrigin-RevId: 52695f24db542428e46b9551e5fca976372d768d