[rdct, rd-necropolis] IJPL-198448: Reuse opened cache whenever possible
If the frontend is reopened without stopping the client, it will try to register the remote caches on the backend. However, if the backend has not been stopped, it will leave the caches opened, so reregistration will fail with the exception.

This commit introduces the reusing of the caches, which should fix the exception


(cherry picked from commit 52695f24db542428e46b9551e5fca976372d768d)

IJ-CR-171354

GitOrigin-RevId: 8240f3634de6d189b7f3d684e33f59f4372ea07e