Speed up API check test
Suspending lazy allows finding recursion, but in an unlucky case it works extremely slowly.
Instead, (1) we can check that there's no cycle in the module graph manually,
and (2) we can consume modules to api machinery in the topological order.

Also, we need to disable recursion check for suspending lazy. Under @Internal ATM.

GitOrigin-RevId: 8bfe8dd9314eead44242dc4d37981b6f1eb05140