[terminal] IJPL-178962 Use absolute values in HighlightTask
Storing TerminalLine / TerminalOffset there was a design
mistake, as they get outdated as soon as the model changes,
and one has to remember that only the absolute values
remain valid.

Instead of trying to remember it, let's only store absolute
values there and convert only when it's known that the model
doesn't change, either because we're doing it in the model
updating coroutine or because the model is frozen.

This fixes the bug when the min(old, new) construct
in the afterContentChanged callback would return
the old value and therefore would store the wrong
relative line, belonging to a previously frozen model
that's no longer valid.

GitOrigin-RevId: 7382d8242aeac384f7b457b91d1a98ed4c9a9fa0