[JEWEL-949] Fix interaction of links and selection in Markdown
This fixes the issue that is present when some selectable `Markdown`
(i.e., when the content is wrapped in a `SelectionContainer`) also has
one or more links. In that case, the `clickable` added to the `Text`
composables used to render paragraphs steals the drag events and makes
it impossible to start selecting text from said `Text` components.

This in turn has a wide-ranging impact in that it makes the
`onTextClick` parameter unused throughout the Markdown APIs. Turns out
that the original use case (making clickable Markdown-based components
such as checkboxes) would be better served by a custom block renderer.
So in this change I removed the `clickable` modifier and `onTextClick`
from all Markdown rendering APIs.

Given this is a large change, even if this is an experimental API, I
decided to still keep the old APIs around for backwards compatibility.
I also took the opportunity to rename the `render` APIs to names that
are more descriptive and respect Compose naming conventions.

Then, I fixed a subtle bug where disabled fenced code blocks would be
more "transparent" than indented ones (we applied `alpha` twice!).

Lastly, I have added a bunch of KDoc around, and cleaned up a lot of
duplication in the `Markdown`/`LazyMarkdown` code, also making sure a
change in `selectable` doesn't lose state.